% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_temp_standardization.R
\name{flux_temp_standardization}
\alias{flux_temp_standardization}
\title{Temperature standardization of gas flux measurements from freshwater systems}
\usage{
flux_temp_standardization(rate_at_source = 2, source_temperature = 15,
  reference_temperature = 20, gas = "CH4")
}
\arguments{
\item{source_temperature}{temperature at which measurements were made}

\item{reference_temperature}{temperature at which measurements are to be standardized}

\item{gas}{gas for which flux was measured}
}
\description{
Temperature standardization of gas flux measurements from freshwater systems
}
\details{
based on DOI: 10.1038/nature13164 ; given an average apparent activation energy, standardizes flux measurements to a reference temperature
}
\examples{
url <- "https://media.nature.com/original/nature-assets/nature/journal/v507/n7493/source_data/nature13164-f2.xlsx"
destfile <- "nature13164_f2.xlsx"
curl::curl_download(url, destfile)
nature13164_f2 <- read_excel(destfile,na="see notes")
nature13164_f2 <- cbind(nature13164_f2, flux_temp_standardization(rate_at_source=nature13164_f2$flux,source_temperature=nature13164_f2$temp, reference_temperature=0))
require(ggplot2)
qplot(data=nature13164_f2,
x=temp, y=log(rate_std_temp),colour=site, facets=.~ecosystem.type)+theme(legend.position="none")+geom_smooth(method="lm",se=F)
}
